Issue 72 / #996: relocate session_glue.rs inline tests to session_glue_tests.rs
session_glue.rs went from 4,800 → 1,237 LOC by lifting the 3,566-line
`#[cfg(test)] mod tests { ... }` block (L1235-4800) into a sibling
session_glue_tests.rs file, loaded via `#[path = "session_glue_tests.rs"]
mod tests;` appended to session_glue.rs.

Pure file-organization split — no test bodies modified, no production
code touched. The test module's `super::*` continues to resolve into
session_glue.rs because `#[path]` only redirects the file location,
not the module's parent.

This is Step 1 of the "Session Glue Deconstruction Roadmap":
- ✅ Issue 72 / #996: isolate glue tests (this PR)
- ⏳ Issue 70 / #994: extract session-key transformation logic
- ⏳ Issue 71 / #995: modularize HA promotion + materialization

LOC distribution after:
- session_glue.rs:        1,237 (was 4,800)
- session_glue_tests.rs:  3,569 (new — same content, dedented one level)

Same playbook as the earlier test-relocation PRs:
- #1050 frame test split (6,435 → 1,995)
- #1051 session test split (3,108 → 1,355)
- #1052 filter test split (2,278 → 1,419)
- #1053 main test split (2,139 → 1,286)

cargo test --release: 865 passed, 0 failed.
